Got UART code to compile. Doesn't appear to work.

+++ code/uart.h
@@ -19,4 +19,4 @@
 */
 
 void uart_init(void);
-uint8_t uart_putchar(char c, FILE *stream);
+int uart_putchar(char c, FILE *stream);
